+++ cmd/gopjrt_installer/darwin.go
@@ -1,4 +1,4 @@
-//go:build (linux && amd64) || all
+//go:build (darwin && arm64) || all
 
 package main
 
@@ -6,34 +6,31 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"runtime"
 	"strings"
 
 	"github.com/pkg/errors"
 )
 
-const AmazonLinux = "amazonlinux"
-
 func init() {
-	for _, plugin := range []string{"linux", AmazonLinux} {
-		pluginInstallers[plugin] = LinuxInstall
-		pluginValidators[plugin] = LinuxValidateVersion
+	for _, plugin := range []string{"darwin"} {
+		pluginInstallers[plugin] = DarwinInstall
+		pluginValidators[plugin] = DarwinValidateVersion
 	}
-	pluginValues = append(pluginValues, "linux", AmazonLinux)
-	pluginDescriptions = append(pluginDescriptions,
-		"CPU PJRT (Linux/amd64)",
-		"CPU PJRT (AmazonLinux/amd64, older libc)")
-	pluginPriorities = append(pluginPriorities, 0, 1)
-	installPathSuggestions = append(installPathSuggestions, "/usr/local/", "~/.local")
+	pluginValues = append(pluginValues, "darwin")
+	pluginDescriptions = append(pluginDescriptions, "CPU PJRT (darwin/arm64)")
+	pluginPriorities = append(pluginPriorities, 3)
+	installPathSuggestions = append(installPathSuggestions, "/usr/local/", "~/Library/Application Support/GoMLX")
 }
 
 // LinuxValidateVersion checks whether the linux version selected by "-version" exists.
-func LinuxValidateVersion(plugin, version string) error {
+func DarwinValidateVersion(plugin, version string) error {
 	// "latest" is always valid.
 	if version == "latest" {
 		return nil
 	}
 
-	_, err := LinuxGetDownloadURL(plugin, version)
+	_, err := DarwinGetDownloadURL(plugin, version)
 	if err != nil {
 		return errors.WithMessagef(err, "can't fetch PJRT plugin from Gopjrt version %q, see "+
 			"https://github.com/gomlx/gopjrt/releases for a list of release versions to choose from", version)
@@ -41,8 +38,8 @@ func LinuxValidateVersion(plugin, version string) error {
 	return err
 }
 
-// LinuxGetDownloadURL returns the download URL for the given version and plugin.
-func LinuxGetDownloadURL(plugin, version string) (url string, err error) {
+// DarwinGetDownloadURL returns the download URL for the given version and plugin.
+func DarwinGetDownloadURL(plugin, version string) (url string, err error) {
 	var assets []string
 	assets, err = GitHubDownloadReleaseAssets(version)
 	if err != nil {
@@ -51,10 +48,8 @@ func LinuxGetDownloadURL(plugin, version string) (url string, err error) {
 
 	var wantAsset string
 	switch plugin {
-	case "linux":
-		wantAsset = "gomlx_xlabuilder_linux_amd64.tar.gz"
-	case AmazonLinux:
-		wantAsset = "gomlx_xlabuilder_linux_amd64_amazonlinux.tar.gz"
+	case "darwin":
+		wantAsset = "gopjrt_cpu_darwin_arm64.tar.gz"
 	default:
 		err = errors.Errorf("version validation not implemented for plugin %q in version %s", plugin, version)
 		return
@@ -67,8 +62,8 @@ func LinuxGetDownloadURL(plugin, version string) (url string, err error) {
 	return "", errors.Errorf("Plugin %q version %q doesn't seem to have the required asset (%q) -- assets found: %v", plugin, version, wantAsset, assets)
 }
 
-// LinuxInstall the assets on the target directory.
-func LinuxInstall(plugin, version, installPath string) error {
+// DarwinInstall the assets on the target directory.
+func DarwinInstall(plugin, version, installPath string) error {
 	var err error
 	if version == "latest" || version == "" {
 		version, err = GitHubGetLatestVersion()
@@ -76,7 +71,7 @@ func LinuxInstall(plugin, version, installPath string) error {
 			return err
 		}
 	}
-	assetURL, err := LinuxGetDownloadURL(plugin, version)
+	assetURL, err := DarwinGetDownloadURL(plugin, version)
 	if err != nil {
 		return err
 	}
@@ -84,6 +79,13 @@ func LinuxInstall(plugin, version, installPath string) error {
 
 	// Create the target directory.
 	installPath = ReplaceTildeInDir(installPath)
+	if strings.Contains(installPath, "Application Support/GoMLX") {
+		// Subdirectory in users Application Support directory is uppercased.
+		installPath = filepath.Join(installPath, "PJRT")
+	} else {
+		// E.g.: installPath = "/usr/local" -> installPath = "/usr/local/lib/gomlx/pjrt"
+		installPath = filepath.Join(installPath, "lib", "gomlx", "pjrt")
+	}
 	if err := os.MkdirAll(installPath, 0755); err != nil {
 		return errors.Wrap(err, "failed to create install directory")
 	}
@@ -112,13 +114,6 @@ func LinuxInstall(plugin, version, installPath string) error {
 		fmt.Printf("  - %s\n", file)
 	}
 
-	// Remove older version using dynamically linked library
-	oldLib := filepath.Join(installPath, "lib/libgomlx_xlabuilder.so")
-	if err := os.Remove(oldLib); err != nil && !os.IsNotExist(err) {
-		ReportError(err)
-	}
-
-	fmt.Printf("\n✅ Installed Gopjrt %s libraries and \"cpu\" PJRT to %s (%s platform)\n\n", version, installPath, plugin)
-
+	fmt.Printf("\n✅ Installed Gopjrt %s \"cpu\" PJRT to %s (%s/%s)\n\n", version, installPath, runtime.GOOS, runtime.GOARCH)
 	return nil
 }